pyproject.toml
[tool.poetry]
name = "online-conference-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
alembic = "^1.13.0"
pytest = "^7.4.3"
asyncpg = "^0.29.0"
pydantic-settings = "^2.1.0"
black = "^23.11.0"
isort = "^5.12.0"
uvicorn = "^0.24.0.post1"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
